+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * `useLineHoldings` hook usage from stripes-acq-components. Refs UIOR-795.
 * Order Tab - Implement MCL Next/Previous pagination. Refs UIOR-782.
 * Location (holding) field adjustments for None or Instance Create inventory setting. Refs UIOR-781.
+* Order lines Tab - Implement MCL Next/Previous pagination. Refs UIOR-783.
 
 ## [2.4.2](https://github.com/folio-org/ui-orders/tree/v2.4.2) (2021-09-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v2.4.1...v2.4.2)